Sync firewall zones colors
In recent versions of LuCI the colors of the firewall zones are
generated by JavaScript code depending on the network name.
In netports code, the zone colors are generated by the old Lua
code from the luci-compat package. As a result, the colors of zones
in the netports table and LuCI network settings are different.

This commit fixes this problem in a such a way that a new JavaScript
code is used to get the colors of the firewall zones.

Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>